created input directory, moved mouse/keyboard.hpp
