+++ src/platform/win32_keyboard.hpp
@@ -4,7 +4,7 @@
 #define NOMINMAX
 #include <windows.h>
 
-#include "core/keyboard.hpp"
+#include "input/keyboard.hpp"
 
 namespace rp::input {
   Keyboard::Key translateWin32KeyCode(WPARAM win32_key_code, LPARAM flags);
